split out tagging to a new workflow step
