+++ spec/robots/dor_repo/gis_assembly/generate_tag_spec.rb
@@ -2,16 +2,10 @@
 
 require 'spec_helper'
 
-RSpec.describe Robots::DorRepo::GisAssembly::GenerateGeoMetadata do
+RSpec.describe Robots::DorRepo::GisAssembly::GenerateTag do
   let(:workflow_client) { instance_double(Dor::Workflow::Client) }
   let(:robot) { described_class.new }
   let(:druid) { 'druid:bb000dd1111' }
-  let(:staging_dir) { File.join(fixture_dir, 'stage', 'bb000dd1111', 'temp') }
-  let(:iso19139_file) { File.join(staging_dir, 'WATER_BODY-iso19139.xml') }
-  let(:iso19139_ng_xml) { Nokogiri::XML(File.read(iso19139_file)) }
-  let(:iso19110_file) { File.join(staging_dir, 'WATER_BODY-iso19110.xml') }
-  let(:iso19110_ng_xml) { Nokogiri::XML(File.read(iso19110_file)) }
-  let(:purl) { 'https://purl.stanford.edu/bb000dd1111' }
 
   before do
     allow(LyberCore::WorkflowClientFactory).to receive(:build).and_return(workflow_client)
@@ -19,7 +13,10 @@ RSpec.describe Robots::DorRepo::GisAssembly::GenerateGeoMetadata do
 
   describe '#perform' do
     let(:object_client) do
-      instance_double(Dor::Services::Client::Object, find: cocina, update: true)
+      instance_double(Dor::Services::Client::Object, find: cocina, update: true, administrative_tags: fake_tags_client)
+    end
+    let(:fake_tags_client) do
+      instance_double(Dor::Services::Client::AdministrativeTags, list: [], create: nil)
     end
     let(:cocina) do
       Cocina::Models::DRO.new(externalIdentifier: 'druid:bb000dd1111',
@@ -42,17 +39,9 @@ RSpec.describe Robots::DorRepo::GisAssembly::GenerateGeoMetadata do
       allow(Dor::Services::Client).to receive(:object).and_return(object_client)
     end
 
-    it 'updates the object' do
+    it 'updates the object with a new tag' do
       test_perform(robot, druid)
-      expect(object_client).to have_received(:update).with(params: Cocina::Models::DRO)
-    end
-  end
-
-  describe '#geo_metadata_rdf_xml' do
-    it 'runs without error' do
-      expect do
-        robot.send(:geo_metadata_rdf_xml, iso19139_ng_xml, iso19110_ng_xml, purl)
-      end.not_to raise_error
+      expect(fake_tags_client).to have_received(:create).once.with(tags: ['Dataset : GIS'])
     end
   end
 end
